+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleaased:
 
 * fix: Reduce idle CPU use
+* fix: Put a timeout on yank command
 
 ## 0.1.7 - 2021-03-28